Query specific variants from specific GWAS

Every rsid is searched for against each requested GWAS id. To get a list of available GWAS ids, or to find their meta data, use gwasinfo. Can request LD proxies for instances when the requested rsid is not present in a particular GWAS dataset. This currently only uses an LD reference panel composed of Europeans in 1000 genomes version 3. It is also restricted to biallelic single nucleotide polymorphisms (no indels), with European MAF > 0.01.

associations(
  variants,
  id,
  proxies = 1,
  r2 = 0.8,
  align_alleles = 1,
  palindromes = 1,
  maf_threshold = 0.3,
  access_token = check_access_token()
)

Arguments

variants

Array of variants e.g. c("rs234", "7:105561135-105563135")

id

Array of GWAS studies to query. See gwasinfo for available studies

proxies

0 or (default) 1 - indicating whether to look for proxies

r2

Minimum proxy LD rsq value. Default=0.8

align_alleles

Try to align tag alleles to target alleles (if proxies = 1). 1 = yes (default), 0 = no

palindromes

Allow palindromic SNPs (if proxies = 1). 1 = yes (default), 0 = no

maf_threshold

MAF threshold to try to infer palindromic SNPs. Default = 0.3.

access_token

Google OAuth2 access token. Used to authenticate level of access to data. By default, checks if already authenticated through get_access_token and if not then does not perform authentication

Value

Dataframe